MSS_Redim_namedRange expansion
The MSS_Redim_namedRange function can be called to stretch the name till the first empty cell instead of the expansing it tothe last non non empty cells in the startingpoint's row and column.  The purpose of this change is that prior to it, it wasn't possible to store multiple blocks underneath or next to each other.